+++ CHANGELOG.md
@@ -7,16 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased] yyyy-mm-dd
 
-### Changed
+### Added
+
+- Issue #4: Immutable geometry, routes, trains
+- Issue #21: Add double switch
+- Issue #28: Simulator with rx4j### Changed
+
+### Changed 
 
 - Issue #3: Stop at platform for loading
-- Issue #4: Immutable geometry, routes, trains
 - Issue #6: Lock signals when train changes section
 - Issue #7: Approaching train state
 - Issue #8: Braking train state
 - Issue #9: Exiting train state
 - Issue #16  Add drawing geometry
 - Issue #18: Change section discovery to handle crossing points
-- Issue #21: Add double switch
 - Issue #23: Add map panel
 - Issue #25: Add station panel